+++ docs/script/search_index.js
@@ -168,12 +168,6 @@ window.esdocSearchIndex = [
     "function"
   ],
   [
-    "formiojs/src/components/day/day.js~daycomponent",
-    "class/src/components/day/Day.js~DayComponent.html",
-    "<span>DayComponent</span> <span class=\"search-result-import-path\">formiojs/src/components/day/Day.js</span>",
-    "class"
-  ],
-  [
     "formiojs/src/displays/displays.js~displays",
     "class/src/displays/Displays.js~Displays.html",
     "<span>Displays</span> <span class=\"search-result-import-path\">formiojs/src/displays/Displays.js</span>",
@@ -546,12 +540,6 @@ window.esdocSearchIndex = [
     "function"
   ],
   [
-    "formiojs/src/components/textfield/textfield.js~textfieldcomponent",
-    "class/src/components/textfield/TextField.js~TextFieldComponent.html",
-    "<span>TextFieldComponent</span> <span class=\"search-result-import-path\">formiojs/src/components/textfield/TextField.js</span>",
-    "class"
-  ],
-  [
     "formiojs/src/components/time/time.form.js~time",
     "function/index.html#static-function-Time",
     "<span>Time</span> <span class=\"search-result-import-path\">formiojs/src/components/time/Time.form.js</span>",
@@ -657,6 +645,12 @@ window.esdocSearchIndex = [
     "formiojs/src/utils/formutils.js~eachcomponent",
     "function/index.html#static-function-eachComponent",
     "<span>eachComponent</span> <span class=\"search-result-import-path\">formiojs/src/utils/formUtils.js</span>",
+    "function"
+  ],
+  [
+    "formiojs/src/embed.js~embed",
+    "function/index.html#static-function-embed",
+    "<span>embed</span> <span class=\"search-result-import-path\">formiojs/src/Embed.js</span>",
     "function"
   ],
   [
@@ -1096,6 +1090,12 @@ window.esdocSearchIndex = [
     "https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/undefined",
     "src/.external-ecmascript.js~undefined",
     "external"
+  ],
+  [
+    "src/embed.js",
+    "file/src/Embed.js.html",
+    "src/Embed.js",
+    "file"
   ],
   [
     "src/eventemitter.js",
@@ -3858,288 +3858,6 @@ window.esdocSearchIndex = [
     "file"
   ],
   [
-    "src/components/day/day.js",
-    "file/src/components/day/Day.js.html",
-    "src/components/day/Day.js",
-    "file"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#_days",
-    "class/src/components/day/Day.js~DayComponent.html#instance-member-_days",
-    "src/components/day/Day.js~DayComponent#_days",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#_months",
-    "class/src/components/day/Day.js~DayComponent.html#instance-member-_months",
-    "src/components/day/Day.js~DayComponent#_months",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#_years",
-    "class/src/components/day/Day.js~DayComponent.html#instance-member-_years",
-    "src/components/day/Day.js~DayComponent#_years",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#attach",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-attach",
-    "src/components/day/Day.js~DayComponent#attach",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#date",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-date",
-    "src/components/day/Day.js~DayComponent#date",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#dayfirst",
-    "class/src/components/day/Day.js~DayComponent.html#instance-member-dayFirst",
-    "src/components/day/Day.js~DayComponent#dayFirst",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#dayrequired",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-dayRequired",
-    "src/components/day/Day.js~DayComponent#dayRequired",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#days",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-days",
-    "src/components/day/Day.js~DayComponent#days",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#defaultschema",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-defaultSchema",
-    "src/components/day/Day.js~DayComponent#defaultSchema",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#disabled",
-    "class/src/components/day/Day.js~DayComponent.html#instance-set-disabled",
-    "src/components/day/Day.js~DayComponent#disabled",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#emptyvalue",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-emptyValue",
-    "src/components/day/Day.js~DayComponent#emptyValue",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#focus",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-focus",
-    "src/components/day/Day.js~DayComponent#focus",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#format",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-format",
-    "src/components/day/Day.js~DayComponent#format",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#getdate",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-getDate",
-    "src/components/day/Day.js~DayComponent#getDate",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#getfieldvalue",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-getFieldValue",
-    "src/components/day/Day.js~DayComponent#getFieldValue",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#getvalidationformat",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-getValidationFormat",
-    "src/components/day/Day.js~DayComponent#getValidationFormat",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#getvalue",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-getValue",
-    "src/components/day/Day.js~DayComponent#getValue",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#getvalueasstring",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-getValueAsString",
-    "src/components/day/Day.js~DayComponent#getValueAsString",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#getvalueat",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-getValueAt",
-    "src/components/day/Day.js~DayComponent#getValueAt",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#init",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-init",
-    "src/components/day/Day.js~DayComponent#init",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#inputdefinition",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-inputDefinition",
-    "src/components/day/Day.js~DayComponent#inputDefinition",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#inputinfo",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-inputInfo",
-    "src/components/day/Day.js~DayComponent#inputInfo",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#ispartialday",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-isPartialDay",
-    "src/components/day/Day.js~DayComponent#isPartialDay",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#monthrequired",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-monthRequired",
-    "src/components/day/Day.js~DayComponent#monthRequired",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#months",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-months",
-    "src/components/day/Day.js~DayComponent#months",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#normalizeminmaxdates",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-normalizeMinMaxDates",
-    "src/components/day/Day.js~DayComponent#normalizeMinMaxDates",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#normalizevalue",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-normalizeValue",
-    "src/components/day/Day.js~DayComponent#normalizeValue",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#parts",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-parts",
-    "src/components/day/Day.js~DayComponent#parts",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#removeinputerror",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-removeInputError",
-    "src/components/day/Day.js~DayComponent#removeInputError",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#render",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-render",
-    "src/components/day/Day.js~DayComponent#render",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#renderfield",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-renderField",
-    "src/components/day/Day.js~DayComponent#renderField",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#selectdefinition",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-selectDefinition",
-    "src/components/day/Day.js~DayComponent#selectDefinition",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#seterrorclasses",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-setErrorClasses",
-    "src/components/day/Day.js~DayComponent#setErrorClasses",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#setvalueat",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-setValueAt",
-    "src/components/day/Day.js~DayComponent#setValueAt",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#shoulddisabled",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-shouldDisabled",
-    "src/components/day/Day.js~DayComponent#shouldDisabled",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#showday",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-showDay",
-    "src/components/day/Day.js~DayComponent#showDay",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#showmonth",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-showMonth",
-    "src/components/day/Day.js~DayComponent#showMonth",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#showyear",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-showYear",
-    "src/components/day/Day.js~DayComponent#showYear",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#validaterequired",
-    "class/src/components/day/Day.js~DayComponent.html#instance-method-validateRequired",
-    "src/components/day/Day.js~DayComponent#validateRequired",
-    "method"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#validationvalue",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-validationValue",
-    "src/components/day/Day.js~DayComponent#validationValue",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#validators",
-    "class/src/components/day/Day.js~DayComponent.html#instance-member-validators",
-    "src/components/day/Day.js~DayComponent#validators",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#valuemask",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-valueMask",
-    "src/components/day/Day.js~DayComponent#valueMask",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#yearrequired",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-yearRequired",
-    "src/components/day/Day.js~DayComponent#yearRequired",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent#years",
-    "class/src/components/day/Day.js~DayComponent.html#instance-get-years",
-    "src/components/day/Day.js~DayComponent#years",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent.builderinfo",
-    "class/src/components/day/Day.js~DayComponent.html#static-get-builderInfo",
-    "src/components/day/Day.js~DayComponent.builderInfo",
-    "member"
-  ],
-  [
-    "src/components/day/day.js~daycomponent.schema",
-    "class/src/components/day/Day.js~DayComponent.html#static-method-schema",
-    "src/components/day/Day.js~DayComponent.schema",
-    "method"
-  ],
-  [
     "src/components/day/day.unit.js",
     "file/src/components/day/Day.unit.js.html",
     "src/components/day/Day.unit.js",
@@ -5646,6 +5364,24 @@ window.esdocSearchIndex = [
     "file"
   ],
   [
+    "src/components/select/fixtures/comp10.js",
+    "file/src/components/select/fixtures/comp10.js.html",
+    "src/components/select/fixtures/comp10.js",
+    "file"
+  ],
+  [
+    "src/components/select/fixtures/comp11.js",
+    "file/src/components/select/fixtures/comp11.js.html",
+    "src/components/select/fixtures/comp11.js",
+    "file"
+  ],
+  [
+    "src/components/select/fixtures/comp12.js",
+    "file/src/components/select/fixtures/comp12.js.html",
+    "src/components/select/fixtures/comp12.js",
+    "file"
+  ],
+  [
     "src/components/select/fixtures/comp2.js",
     "file/src/components/select/fixtures/comp2.js.html",
     "src/components/select/fixtures/comp2.js",
@@ -5673,6 +5409,24 @@ window.esdocSearchIndex = [
     "src/components/select/fixtures/comp6.js",
     "file/src/components/select/fixtures/comp6.js.html",
     "src/components/select/fixtures/comp6.js",
+    "file"
+  ],
+  [
+    "src/components/select/fixtures/comp7.js",
+    "file/src/components/select/fixtures/comp7.js.html",
+    "src/components/select/fixtures/comp7.js",
+    "file"
+  ],
+  [
+    "src/components/select/fixtures/comp8.js",
+    "file/src/components/select/fixtures/comp8.js.html",
+    "src/components/select/fixtures/comp8.js",
+    "file"
+  ],
+  [
+    "src/components/select/fixtures/comp9.js",
+    "file/src/components/select/fixtures/comp9.js.html",
+    "src/components/select/fixtures/comp9.js",
     "file"
   ],
   [
@@ -6478,72 +6232,6 @@ window.esdocSearchIndex = [
     "file/src/components/textfield/TextField.form.js.html",
     "src/components/textfield/TextField.form.js",
     "file"
-  ],
-  [
-    "src/components/textfield/textfield.js",
-    "file/src/components/textfield/TextField.js.html",
-    "src/components/textfield/TextField.js",
-    "file"
-  ],
-  [
-    "src/components/textfield/textfield.js~textfieldcomponent#defaultschema",
-    "class/src/components/textfield/TextField.js~TextFieldComponent.html#instance-get-defaultSchema",
-    "src/components/textfield/TextField.js~TextFieldComponent#defaultSchema",
-    "member"
-  ],
-  [
-    "src/components/textfield/textfield.js~textfieldcomponent#emptyvalue",
-    "class/src/components/textfield/TextField.js~TextFieldComponent.html#instance-get-emptyValue",
-    "src/components/textfield/TextField.js~TextFieldComponent#emptyValue",
-    "member"
-  ],
-  [
-    "src/components/textfield/textfield.js~textfieldcomponent#getvalueat",
-    "class/src/components/textfield/TextField.js~TextFieldComponent.html#instance-method-getValueAt",
-    "src/components/textfield/TextField.js~TextFieldComponent#getValueAt",
-    "method"
-  ],
-  [
-    "src/components/textfield/textfield.js~textfieldcomponent#inputinfo",
-    "class/src/components/textfield/TextField.js~TextFieldComponent.html#instance-get-inputInfo",
-    "src/components/textfield/TextField.js~TextFieldComponent#inputInfo",
-    "member"
-  ],
-  [
-    "src/components/textfield/textfield.js~textfieldcomponent#isempty",
-    "class/src/components/textfield/TextField.js~TextFieldComponent.html#instance-method-isEmpty",
-    "src/components/textfield/TextField.js~TextFieldComponent#isEmpty",
-    "method"
-  ],
-  [
-    "src/components/textfield/textfield.js~textfieldcomponent#maskvalue",
-    "class/src/components/textfield/TextField.js~TextFieldComponent.html#instance-method-maskValue",
-    "src/components/textfield/TextField.js~TextFieldComponent#maskValue",
-    "method"
-  ],
-  [
-    "src/components/textfield/textfield.js~textfieldcomponent#normalizevalue",
-    "class/src/components/textfield/TextField.js~TextFieldComponent.html#instance-method-normalizeValue",
-    "src/components/textfield/TextField.js~TextFieldComponent#normalizeValue",
-    "method"
-  ],
-  [
-    "src/components/textfield/textfield.js~textfieldcomponent#setvalueat",
-    "class/src/components/textfield/TextField.js~TextFieldComponent.html#instance-method-setValueAt",
-    "src/components/textfield/TextField.js~TextFieldComponent#setValueAt",
-    "method"
-  ],
-  [
-    "src/components/textfield/textfield.js~textfieldcomponent.builderinfo",
-    "class/src/components/textfield/TextField.js~TextFieldComponent.html#static-get-builderInfo",
-    "src/components/textfield/TextField.js~TextFieldComponent.builderInfo",
-    "member"
-  ],
-  [
-    "src/components/textfield/textfield.js~textfieldcomponent.schema",
-    "class/src/components/textfield/TextField.js~TextFieldComponent.html#static-method-schema",
-    "src/components/textfield/TextField.js~TextFieldComponent.schema",
-    "method"
   ],
   [
     "src/components/textfield/textfield.unit.js",
@@ -8769,6 +8457,12 @@ window.esdocSearchIndex = [
     "src/templates/index.js",
     "file/src/templates/index.js.html",
     "src/templates/index.js",
+    "file"
+  ],
+  [
+    "src/translations/en.js",
+    "file/src/translations/en.js.html",
+    "src/translations/en.js",
     "file"
   ],
   [