+++ src/Webform.js
@@ -293,25 +293,22 @@ export default class Webform extends NestedDataComponent {
    * @return {Promise}
    */
   set language(lang) {
-    return new NativePromise((resolve, reject) => {
-      this.options.language = lang;
-      if (this.i18next.language === lang) {
-        return resolve();
-      }
-      try {
-        this.i18next.changeLanguage(lang, (err) => {
-          if (err) {
-            return reject(err);
-          }
-          this.redraw();
-          this.emit('languageChanged');
-          resolve();
-        });
-      }
-      catch (err) {
-        return reject(err);
-      }
-    });
+    this.options.language = lang;
+    if (this.i18next.language === lang) {
+      return;
+    }
+    try {
+      this.i18next.changeLanguage(lang, (err) => {
+        if (err) {
+          return;
+        }
+        this.redraw();
+        this.emit('languageChanged');
+      });
+    }
+    catch (err) {
+      return;
+    }
   }
 
   get componentComponents() {
@@ -663,7 +660,7 @@ export default class Webform extends NestedDataComponent {
       this._form = flags?.keepAsReference ? form : _.cloneDeep(form);
 
       if (this.onSetForm) {
-        this.onSetForm(this._form, form);
+        this.onSetForm(_.cloneDeep(this._form), form);
       }
     }
     catch (err) {
@@ -1305,6 +1302,10 @@ export default class Webform extends NestedDataComponent {
       noCheck: true
     });
     this.setAlert('success', `<p>${this.t('complete')}</p>`);
+    // Cancel triggered saveDraft to prevent overriding the submitted state
+    if (this.draftEnabled && this.triggerSaveDraft?.cancel) {
+      this.triggerSaveDraft.cancel();
+    }
     this.emit('submit', submission, saved);
     if (saved) {
       this.emit('submitDone', submission);