+++ src/WebformBuilder.js
@@ -110,14 +110,8 @@ export default class WebformBuilder extends Component {
         }
       }
 
-      // Order the compoennts.
-      if (info.components) {
-        info.componentOrder = Object.keys(info.components)
-          .map(key => info.components[key])
-          .filter(component => component && !component.ignore && !component.ignoreForForm)
-          .sort((a, b) => a.weight - b.weight)
-          .map(component => component.key);
-      }
+      // Order the components.
+      this.orderComponents(info);
     }
 
     this.options.hooks = this.options.hooks || {};
@@ -614,22 +608,22 @@ export default class WebformBuilder extends Component {
     const filterGroupBy = (group, searchValue = '') => {
       const result = _.toPlainObject(group);
       const { subgroups = [], components } = result;
-      const filteredOrder = [];
+      const filteredComponents = [];
 
       for (const key in components) {
         const isMatchedToTitle = components[key].title.toLowerCase().match(searchValue);
         const isMatchedToKey = components[key].key.toLowerCase().match(searchValue);
 
         if (isMatchedToTitle || isMatchedToKey) {
-          filteredOrder.push(components[key].key);
+          filteredComponents.push(components[key]);
         }
       }
 
-      result.componentOrder = filteredOrder;
+      this.orderComponents(result, filteredComponents);
       if (searchValue) {
         result.default = true;
       }
-      if (filteredOrder.length || subgroups.length) {
+      if (result.componentOrder.length || subgroups.length) {
         return result;
       }
       return null;
@@ -674,6 +668,17 @@ export default class WebformBuilder extends Component {
     });
 
     this.updateDragAndDrop();
+  }
+
+  orderComponents(groupInfo, foundComponents) {
+    const components = foundComponents || groupInfo.components;
+    if (components) {
+      groupInfo.componentOrder = Object.keys(components)
+        .map(key => components[key])
+        .filter(component => component && !component.ignore && !component.ignoreForForm)
+        .sort((a, b) => a.weight - b.weight)
+        .map(component => component.key);
+    }
   }
 
   updateDragAndDrop() {
@@ -1045,7 +1050,7 @@ export default class WebformBuilder extends Component {
     // Update the preview.
     if (this.preview) {
       this.preview.form = {
-        components: [_.omit(component, [
+        components: [_.omit({ ...component }, [
           'hidden',
           'conditional',
           'calculateValue',
@@ -1071,7 +1076,7 @@ export default class WebformBuilder extends Component {
       );
 
       if (!defaultChanged) {
-        _.assign(defaultValueComponent.component, _.omit(component, [
+        _.assign(defaultValueComponent.component, _.omit({ ...component }, [
           'key',
           'label',
           'placeholder',
@@ -1085,6 +1090,7 @@ export default class WebformBuilder extends Component {
           'calculateValue',
           'conditional',
           'customConditional',
+          'id'
         ]));
         const parentComponent = defaultValueComponent.parent;
         let tabIndex = -1;