+++ src/validator/Validator.js
@@ -51,6 +51,7 @@ class ValidationChecker {
       onlyAvailableItems: {
         key: 'validate.onlyAvailableItems',
         method: 'validateValueAvailability',
+        hasLabel: true,
         message(component) {
           return component.t(component.errorMessage('valueIsNotAvailable'), {
             field: component.errorLabel,
@@ -155,7 +156,13 @@ class ValidationChecker {
               }
               else if (result) {
                // Only OK if it matches the current submission
-                return resolve(submission._id && (result._id.toString() === submission._id));
+                if (submission._id && (result._id.toString() === submission._id)) {
+                  resolve(true);
+                }
+                else {
+                  component.conflictId = result._id.toString();
+                  return resolve(false);
+                }
               }
               else {
                 return resolve(true);
@@ -376,7 +383,8 @@ class ValidationChecker {
             return total;
           }, 0);
 
-          return count >= min;
+          // Should not be triggered if there is no options selected at all
+          return !count || count >= min;
         }
       },
       maxSelectedCount: {
@@ -417,7 +425,7 @@ class ValidationChecker {
         },
         check(component, setting, value) {
           const minLength = parseInt(setting, 10);
-          if (!minLength || (typeof value !== 'string') || component.isEmpty(value)) {
+          if (!value || !minLength || (typeof value !== 'string') || component.isEmpty(value)) {
             return true;
           }
           return (value.length >= minLength);
@@ -471,7 +479,7 @@ class ValidationChecker {
         },
         check(component, setting, value) {
           const minWords = parseInt(setting, 10);
-          if (!minWords || (typeof value !== 'string')) {
+          if (!minWords || !value || (typeof value !== 'string')) {
             return true;
           }
           return (value.trim().split(/\s+/).length >= minWords);
@@ -650,6 +658,8 @@ class ValidationChecker {
           inputMask = inputMask ? getInputMask(inputMask) : null;
 
           if (value && inputMask && !component.skipMaskValidation) {
+            // If char which is used inside mask placeholder was used in the mask, replace it with space to prevent errors
+            inputMask = inputMask.map((char) => char === component.placeholderChar ? ' ' : char);
             return matchInputMask(value, inputMask);
           }
 
@@ -882,19 +892,28 @@ class ValidationChecker {
     const resultOrPromise = this.checkValidator(component, validator, setting, value, data, index, row, async);
 
     const processResult = result => {
-      return result ? {
-        message: unescapeHTML(_.get(result, 'message', result)),
-        level: _.get(result, 'level') === 'warning' ? 'warning' : 'error',
-        path: getArrayFromComponentPath(component.path || ''),
-        context: {
-          validator: validatorName,
-          hasLabel: validator.hasLabel,
-          setting,
-          key: component.key,
-          label: component.label,
-          value
+      if (result) {
+        const resultData = {
+          message: unescapeHTML(_.get(result, 'message', result)),
+          level: _.get(result, 'level') === 'warning' ? 'warning' : 'error',
+          path: getArrayFromComponentPath(component.path || ''),
+          context: {
+            validator: validatorName,
+            hasLabel: validator.hasLabel,
+            setting,
+            key: component.key,
+            label: component.label,
+            value
+          }
+        };
+        if (validatorName ==='unique' && component.conflictId) {
+          resultData.conflictId = component.conflictId;
         }
-      } : false;
+        return resultData;
+      }
+      else {
+        return false;
+      }
     };
 
     if (async) {